chore:migrated HomeScreen screen into Jetpack compose

+++ app/src/main/java/com/sorabh/grabfood/ui/viewmodel/HomeViewModel.kt
@@ -10,11 +10,10 @@ import com.sorabh.grabfood.domain.usecase.GetRestaurantUseCase
 import com.sorabh.grabfood.util.Constants
 import com.sorabh.grabfood.util.Keys
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -46,26 +45,15 @@ class HomeViewModel @Inject constructor(
         }
     }
 
-    suspend fun insertRestaurant(dish: Dish): Boolean {
-        val result = CoroutineScope(Dispatchers.IO).async {
-            localDBRepository.insertRestaurant(dish)
-            isRestaurantStored(dish.id)
+    fun onFavoriteIconBtnClick(dish: Dish) {
+        viewModelScope.launch {
+            if (localDBRepository.getRestaurant(dish.id) > 0)
+                localDBRepository.deleteRestaurant(dish)
+            else
+                localDBRepository.insertRestaurant(dish)
         }
-        return result.await()
     }
-
-    suspend fun deleteRestaurant(dish: Dish): Boolean {
-        val result = CoroutineScope(Dispatchers.IO).async {
-            localDBRepository.deleteRestaurant(dish)
-            isRestaurantStored(dish.id)
-        }
-        return !result.await()
-    }
-
-    suspend fun isRestaurantStored(id: String): Boolean {
-        val result = CoroutineScope(Dispatchers.IO).async {
-            localDBRepository.getRestaurant(id)
-        }
-        return result.await() > 0
+    fun isDishStored(id: String): Flow<Boolean> = flow {
+        emit(localDBRepository.getRestaurant(id) > 0)
     }
 }
